backend/requirements.txt
pyproject-toml
psycopg2
Django
environs
gunicorn
pandas
djangorestframework
django-filter
Pillow